+++ src/main/java/org/zalando/nakadi/service/EventStreamFactory.java
@@ -2,6 +2,7 @@ package org.zalando.nakadi.service;
 
 import com.codahale.metrics.Meter;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.zalando.nakadi.exceptions.InvalidCursorException;
 import org.zalando.nakadi.exceptions.NakadiException;
@@ -15,17 +16,25 @@ public class EventStreamFactory {
     private final CursorConverter cursorConverter;
     private final EventStreamWriterProvider writerProvider;
     private final BlacklistService blacklistService;
+    private final NakadiKpiPublisher nakadiKpiPublisher;
+    private final String kpiDataStreamedEventType;
+    private final long kpiFrequencyMs;
 
     @Autowired
     public EventStreamFactory(
             final CursorConverter cursorConverter,
             final EventStreamWriterProvider writerProvider,
-            final BlacklistService blacklistService) {
+            final BlacklistService blacklistService,
+            final NakadiKpiPublisher nakadiKpiPublisher,
+            @Value("${nakadi.kpi.event-types.nakadiDataStreamed}") final String kpiDataStreamedEventType,
+            @Value("${nakadi.kpi.config.stream-data-collection-frequency-ms}") final long kpiFrequencyMs) {
         this.cursorConverter = cursorConverter;
         this.writerProvider = writerProvider;
         this.blacklistService = blacklistService;
+        this.nakadiKpiPublisher = nakadiKpiPublisher;
+        this.kpiDataStreamedEventType = kpiDataStreamedEventType;
+        this.kpiFrequencyMs = kpiFrequencyMs;
     }
-
 
     public EventStream createEventStream(final OutputStream outputStream, final EventConsumer eventConsumer,
                                          final EventStreamConfig config, final Meter bytesFlushedMeter)
@@ -37,6 +46,9 @@ public class EventStreamFactory {
                 blacklistService,
                 cursorConverter,
                 bytesFlushedMeter,
-                writerProvider);
+                writerProvider,
+                nakadiKpiPublisher,
+                kpiDataStreamedEventType,
+                kpiFrequencyMs);
     }
 }