+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [2.5.0] - 2017-12-22
+
+### Added
+- Nakadi collects event publishing KPI data 
+- Nakadi collects event streaming KPI data
+
 ## [2.4.2] - 2017-12-21
 
 ### Added